All assets are in "merge assets" folder...
All assets are in "merge assets" folder. Map works okay, but will need to add walls, i think ill just add some rocks to block.,

Big issue is the navigation does not like all the random poligon, but before that it was also acting up. Im not exactly sure why, it may be the code. Gonna check the code later.